Adding support to do real prefetching on libmemcached
When calling mget, the function flushes all the previous results. This blocks
the usual prefetching strategy which consists on doing many calls to mget and
fetch all the results later on. We added an behaviour flag, to avoid this when
we want to do prefetching.

The flag is MEMCACHED_BEHAVIOR_MGET_FLUSH_OLD_RESULTS, and the defaul value is
true. Set this to false if you want prefetch results.