Solved issue with packet lost or reordering on UDP multigets
When receiving large multiget responses, some reordering could lead
to an incorrect assembling of the response, mixing two request_ids
without being properly detected.

The problem is triggered when receiving the first datagram of a response
(udp_header->sequence_number == 0) in the middle of a proccessing of
other UDP response.

The hard part is to properly detect this situation and also handling
packet loss properly. In order to do this, we reset ptr->num_datagrams
at binary_read_one_response to indicate the beginning of a new response.